Move z score logic into its own file
